refactor port logic and path hashing
