+++ src/build-config/default-build-config.js
@@ -1,7 +1,6 @@
 import { CX_22_0 } from "../version";
 import { LANDINGPAGE } from "../design-type";
 import DistFolder from "../dist-folder";
-import * as net from 'net';
 
 /** @typedef {import('./build-config-interface').default} BuildConfigInterface */
 
@@ -30,11 +29,7 @@ export default class DefaultBuildConfig {
   }
 
   get devServerPort() {
-    // return 'auto';
-    let devPort = 9003; 
-    //this._assignPort(devPort).then(port => devPort = port);
-    //console.log("Default port is: /d", devPort);
-    return devPort;
+    return 'auto';
   }
 
   get hashZipFiles() {
@@ -91,111 +86,5 @@ export default class DefaultBuildConfig {
 
   get postcssEnabled() {
     return false;
-  }
-
-  async isPortReachable(port, {host, timeout = 1000} = {}) {
-    if (typeof host !== 'string') {
-      throw new TypeError('Specify a `host`');
-    }
-  
-    const promise = new Promise(((resolve, reject) => {
-      const socket = new net.Socket();
-  
-      const onError = () => {
-        socket.destroy();
-        reject();
-      };
-  
-      socket.setTimeout(timeout);
-      socket.once('error', onError);
-      socket.once('timeout', onError);
-  
-      socket.connect(port, host, () => {
-        socket.end();
-        resolve();
-      });
-    }));
-  
-    try {
-      await promise;
-      return true;
-    } catch {
-      return false;
-    }
-  }
-
-  /**
-   * @param {number} port - The port to assign
-   * @returns {number}
-   * @private
-   */
-  async _assignPort(defaultPort) {
-    let maxPort = 9200;
-    if (defaultPort > maxPort) {
-      return;
-    }
-    let port = defaultPort;
-    const net = require("net");
-
-    let usePort = false;
-    
-   // while (port < maxPort) {
-   //   await(this.isPortReachable(port, { host: "localhost" }))
-   //   .then((isUsable) => {if (isUsable) {
-   //    return port; 
-   //   }});
-   //   console.log("Port:" + port + " , use:" + usePort);
-   //   port+=1;
-   // }
-   // 
-   // return port;
-    console.log("Start listening on %d", port);
-
-    const server = net.createServer();
-
-    // handle connection error
-    server.once("error", (err) => {
-      usePort = false;
-      console.log("Cannot connect to port %d. Error: " + err, port);
-      // port = this._assignPort(port + 1);
-    });
-
-    // connection successful
-    server.on("listening", function () {
-      console.log("Connection to port %d successful", port);
-      server.close();
-      usePort = true;
-    });
-    server.on("connection", function () {
-      console.log("Connection to port %d successful", port);
-      server.close();
-      usePort = true;
-    });
-
-  //  while (!usePort) {
-  //    console.log(await isPortReachable(port, { host: "localhost" }));
-  //    port+=1;
-  //  }
-
-     await(
-       server.listen(port, () => {
-       console.log("Actual port: %d", port);
-       usePort = true;
-       return port;
-     })
-     );
-
-    console.log(server);
-    server.close();
-    // TODO: replace usePort with code to wait on listen call.
-    // socket stuff is async, watch out for inf loops
-    if (usePort) {
-      console.log("Final portNr = &d", port);
-      return port;
-    }
-    else {
-      console.log("DEBUG: ELSE");
-      port = this._assignPort(port + 1);
-    }
   }
 }
