+++ src/path-hash-utility.js
@@ -1,0 +1,24 @@
+String.prototype.hashCode = function() {
+  let hash = 0,
+    i, chr;
+  if (this.length === 0) return hash;
+  for (i = 0; i < this.length; i++) {
+    chr = this.charCodeAt(i);
+    hash = ((hash << 5) - hash) + chr;
+    hash |= 0; // Convert to 32bit integer
+  }
+  return hash;
+}
+
+/**
+ * @param {string} filePath
+ * @returns {number}
+ */
+export function _createPathHash(filePath) {
+  let hash = filePath.toString().hashCode();
+  if (hash < 0) {
+    hash *= -1;
+    hash = hash + '1';
+  }
+  return hash;
+}
